[REFACTOR] : useAuth 커스텀 훅 추가
